Control moved to unix socket
Program now only listens to input from the unix socket. There is no "command line". There isn't a lot of structure to
this interface yet, but the program can be closed by sending a message with body "quit" on this socket using curl. This
is just for testing purposes. A full API now needs to be developed.