commandQueue made abstract
commandQueue now anything implementing a typeclass which encodes "can be interpreted in a Manager".

Moved commandQueue in to Manager

commandQueue now moved in to Manager, since it doesn't really have anything to do with the Control element
(the controller will put commands on the queue but that's all).

Rewrote commandQueue using typeclass

commandQueue is now more abstract, and doesn't need to encodeall of the possible commands. This makes it quite
extensible.

Tidy

Tidy up the Manager file.