Tracing controller
Creating a skeleton controller which just dispatches add commands. Manager monad wasn't the correct
abstraction (since warp does the threading we need anyway), so opted for a Reader monad for the moment.
Requires a bit of faff to get the types to line up because Applications (from wai) aren't polymorphic.