Made new connections fully capable
Now we can start a new connection with no knowledge of an endpoint, and then later connect that endpoint when we know
what we should be connecting to, or we can develop a connection to an address which we've been told about.

Squash of:

Make connect more capable

connect can now deal with "empty" sockets. In fact, non-empty sockets are pointless to stash, so we just add them
immediately

Renamed New

New isn't really new, but rather a response to an offer already being made - renamed to Develop.

Added new command

Creates a socket that can be offered to connect to another endpoint.

Fixes

Message with new connection needed to be dispatched to the Manager, and we needed to return the address of the endpoint
on a connection develop.